TaskView: Remove unecessary asserts & disable array index lint rule
Having array keys as indices is okay iff we don't manipulate the order
of the blocks. We're not doing that here, so it's fine to keep them as
indices for now.

In the event that we will want to manipulate them, we'll have to change
the TaskBlockFragment types to include an id to be used as a key.